+++ PROG_II_gymFIT_integrador/PROG_II_gymFIT__integrador/Persona.h
@@ -7,24 +7,24 @@ class Persona
 private:
 	char _nombre[100];
 	char _apellido[100];
-	int _DNI;
+	long long _DNI;
 	int _edad;
 	Fecha _fechaNacimiento;
 	char _correoElectronico[100];
 public:
 	Persona();
-	Persona(std::string nombre, std::string apellido, int dni, int edad, int dia, int mes, int anio, std::string correoElectronico);
+	Persona(std::string nombre, std::string apellido, long long dni, int edad, int dia, int mes, int anio, std::string correoElectronico);
 	
 	// setters
 	void setNombre(std::string nombre);
 	void setApellido(std::string apellido);
-	void setDNI(int dni);
+	void setDNI(long long dni);
 	void setCorreoElectronico(std::string correoElectronico);
 	
 	// getters
 	std::string getNombre() const;
 	std::string getApellido() const;
-	int getDNI() const;
+	long long getDNI() const;
 	Fecha getFechaNacimiento() const;
 	int getEdad() const;
 	std::string getCorreoElectronico() const;
